+++ history.md
@@ -1,3 +1,12 @@
+## Edge
+
+* Add deprecation warning for `wti --fetch`, `wti -f`. These commands will be deprecated in favour of `wti pull`.
+* Add deprecation warning for `wti --upload`, `wti -u`. These commands will be deprecated in favour of `wti push`.
+* Add -c parameter to specify a configuration file at a custom location.
+* Add -l parameter to specify a specific language file to pull or push (only works with `wti pull` and `wti push`).
+* Add --force parameter to force Web Translate It to send the language files again, regardless if the current 
+  language file version is current (this makes `wti pull` very much slower).
+
 ## Version 1.4.6 /2010-02-04
 
 * Add feedback when using the `wti command`.